test: refatorando arquivos de testes
