+++ test/clientController.delete.test.js
@@ -1,0 +1,4 @@
+const clientController = require('../src/controllers/clientController');
+const clientModel = require('../src/models/clientModel');
+
+jest.mock('../src/models/clientModel');
